# specs/001-product-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SmartInventory API
  version: 1.0.0
  description: |
    Contract for the first SmartInventory API practice covering health status
    and in-memory product management.
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      summary: Get service health
      operationId: getHealth
      responses:
        '200':
          description: Service is available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                default:
                  value:
                    status: UP
                    service: SmartInventory API
                    version: 1.0.0
  /api/products:
    get:
      summary: List products
      operationId: listProducts
      responses:
        '200':
          description: Products returned successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
    post:
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid product input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Product identifier
        schema:
          type: integer
          format: int64
    get:
      summary: Get product by ID
      operationId: getProductById
      responses:
        '200':
          description: Product returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update product by ID
      operationId: updateProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Invalid product input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete product by ID
      operationId: deleteProduct
      responses:
        '204':
          description: Product deleted successfully
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ProductRequest:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          format: double
          exclusiveMinimum: 0
        stock:
          type: integer
          minimum: 0
    ProductResponse:
      type: object
      required:
        - id
        - name
        - description
        - price
        - stock
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        stock:
          type: integer
    HealthResponse:
      type: object
      required:
        - status
        - service
        - version
      properties:
        status:
          type: string
          example: UP
        service:
          type: string
          example: SmartInventory API
        version:
          type: string
          example: 1.0.0
    ErrorResponse:
      type: object
      required:
        - message
        - path
        - timestamp
      properties:
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        errors:
          type: object
          additionalProperties:
            type: string